Script 10

+++ pointers_arrays_strings/100-atoi.c
@@ -10,7 +10,7 @@
 int _atoi(char *s)
 {
 int n = 0;
-int num = 0;
+double num = 0;
 int sign = 1;
 
 	while (s[n] != '\0')
